requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
ape-foundry==0.8.7
ape-infura==0.8.4
ape-solidity==0.8.5
ape-vyper==0.8.7
astroid==3.3.9
asttokens==2.4.1
attrs==24.3.0
base58==1.0.3
bitarray==3.0.0
cached-property==2.0.1
cbor2==5.6.5
cchecksum==0.0.4
certifi==2024.12.14
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
cytoolz==1.0.1
dataclassy==0.11.1
decorator==5.1.1
dill==0.4.0
eip712==0.2.10
eth-account==0.13.4
eth-ape==0.8.24
eth-bloom==3.0.1
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-pydantic-types==0.1.3
eth-rlp==2.1.0
eth-tester==0.12.0b2
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
ethpm-types==0.6.19
evm-trace==0.2.4
evmchains==0.1.3
executing==2.1.0
fonttools==4.57.0
frozenlist==1.5.0
hexbytes==1.2.1
idna==3.10
ijson==3.3.0
importlib_metadata==8.5.0
iniconfig==2.0.0
ipython==8.31.0
isort==6.0.1
jedi==0.19.2
kiwisolver==1.4.8
lark==1.2.2
lazyasd==0.1.4
lru-dict==1.3.0
markdown-it-py==3.0.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
morphys==1.0
msgspec==0.19.0
multidict==6.1.0
numpy==1.26.4
packaging==23.2
pandas==2.2.3
parsimonious==0.10.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cid==0.3.0
py-ecc==7.0.1
py-evm==0.10.1b2
py-geth==5.1.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==0.2.3
py-solc-x==2.0.3
pycryptodome==3.21.0
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.18.0
pylint==3.3.6
pyparsing==3.2.3
pytest==8.3.4
python-baseconv==1.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rlp==4.0.1
safe-pysha3==1.0.4
semantic-version==2.10.0
six==1.17.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
stack-data==0.6.3
tomlkit==0.13.2
toolz==1.0.0
tqdm==4.67.1
traitlets==5.14.3
trie==3.0.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
varint==1.0.2
vvm==0.2.1
vyper==0.4.1
watchdog==3.0.0
wcwidth==0.2.13
web3==7.6.1
websockets==13.1
yarl==1.18.3
zipp==3.21.0
